Extract Rails path manipulation into separate fixture helper
