+++ test/support/fixture_helper.rb
@@ -1,8 +1,6 @@
 # typed: ignore
 # frozen_string_literal: true
 
-require "rails_test_helper"
-
 module FixtureHelper
   TEMP_FIXTURE_DIR = ROOT.join("tmp", "fixtures").to_s
   FileUtils.mkdir_p(TEMP_FIXTURE_DIR)
@@ -14,11 +12,9 @@ module FixtureHelper
 
   def around
     old_working_dir = Dir.pwd
-    cache_rails_paths
 
     yield
 
-    restore_rails_paths
     Dir.chdir(old_working_dir)
     FileUtils.remove_entry(@app_dir, true) if defined? @app_dir
   end
@@ -26,29 +22,6 @@ module FixtureHelper
   def copy_template(template)
     copy_dir("test/fixtures/#{template}")
     Dir.chdir(app_dir)
-
-    case template
-    when :minimal
-      rails_autoload_paths(paths_to("/components/sales/app/models"))
-      rails_eager_load_paths([])
-      rails_autoload_once_paths([])
-      rails_root_path(app_dir)
-    when :skeleton
-      rails_autoload_paths(paths_to("components/sales/app/models"))
-      rails_eager_load_paths(paths_to("components/platform/app/models"))
-      rails_autoload_once_paths(
-        paths_to(
-          "components/timeline/app/models",
-          "components/timeline/app/models/concerns",
-          "vendor/cache/gems/example/models",
-        )
-      )
-      rails_root_path(app_dir)
-      # make sure PrivateThing.constantize succeeds to pass the privacy validity check
-      require "fixtures/skeleton/components/timeline/app/models/private_thing.rb"
-    else
-      raise "Unknown fixture template #{template}"
-    end
   end
 
   def copy_dir(path)
@@ -88,36 +61,6 @@ module FixtureHelper
   end
 
   private
-
-  def rails_autoload_paths(paths)
-    Rails.application.config.autoload_paths = paths
-  end
-
-  def rails_eager_load_paths(paths)
-    Rails.application.config.eager_load_paths = paths
-  end
-
-  def rails_autoload_once_paths(paths)
-    Rails.application.config.autoload_once_paths = paths
-  end
-
-  def rails_root_path(path)
-    Rails.application.config.root = path
-  end
-
-  def cache_rails_paths
-    @rails_autoload_paths = Rails.application.config.autoload_paths
-    @rails_eager_load_paths = Rails.application.config.eager_load_paths
-    @rails_autoload_once_paths = Rails.application.config.autoload_once_paths
-    @rails_root_path = Rails.application.config.root
-  end
-
-  def restore_rails_paths
-    rails_autoload_paths(@rails_autoload_paths)
-    rails_eager_load_paths(@rails_eager_load_paths)
-    rails_autoload_once_paths(@rails_autoload_once_paths)
-    rails_root_path(@rails_root_path)
-  end
 
   def recursive_merge!(hash, other_hash)
     hash.merge!(other_hash) do |_, old_value, new_value|
